pyproject.toml
[project]
name = "yofardev-captioner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Pillow",
    "torch",
    "huggingface_hub",
    "transformers",
    "mistralai",
    "python-dotenv",
    "openai"
]
